requirements.txt
pydantic
requests
PyYAML>=6.0.3